refactor(perf): split /start + /community into server shell + motion island
Mirrors W17-W22 pattern. 7th + 8th top-of-funnel split.

- app/start/page.tsx: 346 -> 5 lines (-341), motion moved to components/start/StartShell.tsx
- app/community/page.tsx: 404 -> 5 lines (-399), motion moved to components/community/CommunityShell.tsx
- No FAQ extraction needed (neither page had FAQ arrays)
- Metadata stays in app/community/layout.tsx (unchanged)
- Server pages now ship zero client JS for static-shell render
- npx tsc --noEmit: clean
